+++ src/main/kotlin/day09/Day09.kt
@@ -3,11 +3,11 @@ package day09
 import Day
 
 class Day09(private val input: String) : Day() {
-    override fun part1(): Int {
+    override fun part1(): Long {
         return 0
     }
 
-    override fun part2(): Int {
+    override fun part2(): Long {
         return 0
     }
 }
